pyproject.toml
[project]
name = "mcp-server-weather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.3",
    "resonate-sdk>=0.5.4",
]
